[rts] Turn Result into a protobuf
Turn Result and Threshold structs into a protobuf, for serialization
stability.

Bug: 1112125
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2681152
Commit-Queue: Nodir Turakulov <nodir@chromium.org>
Cr-Commit-Position: refs/heads/master@{#38321}

+++ go/src/infra/rts/presubmit/eval/run_test.go
@@ -16,7 +16,7 @@ func TestScoreString(t *testing.T) {
 
 	Convey(`ScoreString`, t, func() {
 		Convey("NaN", func() {
-			So(scoreString(math.NaN()), ShouldEqual, "?")
+			So(scoreString(float32(math.NaN())), ShouldEqual, "?")
 		})
 		Convey("0%", func() {
 			So(scoreString(0), ShouldEqual, "0.00%")
